accept window id as argument
Instead of using the window which has currently the focus, accept any window id
you get as first argument. The window id has to be in hex format with leading
'0x', otherwise it has to be a decimal number. If no argument is given, the
window which has the focus is used.