implement registrar dbus service
I added the real functionality to the dbus service and also found the real
reason why the kded module 'appmenu' crashes: If the window isn't registered,
GetMenuForWindow returns an empty service name and an empty object path. An
assert in libdbus checks if the object path is valid and an empty path fails.
This aborts the kded module.

qmenu_registrar returns a dbus error when the window isn't registered. This
error should be handled gracefully in qmenu_hud. At the moment nothing is
shown.