added experimental gtk support
Unity doesn't use appmenu for gtk anymore. The unity gtk module implements a
gtk specific menu handling. The dbus interface is even an implementation
detail, so it can break any minute.

With gtk one can also not use XGetInputFocus() as gtk sets focus to some child
window, not the top level window which we need. It's now reading
_NET_ACTIVE_WINDOW from the default root window assuming the window manager
sets this.